+++ srcs/drawer/hit_cones.c
@@ -6,7 +6,7 @@
 /*   By: dshatilo <dshatilo@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/19 14:25:08 by dshatilo          #+#    #+#             */
-/*   Updated: 2024/07/19 23:26:15 by dshatilo         ###   ########.fr       */
+/*   Updated: 2024/07/29 15:03:20 by dshatilo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ static void	hit_curved_surface_cone(t_conic *cone, t_ray *ray, t_hp *hp)
 	t_vec3		norm;
 
 	hp->type = OBJ_NONE;
-	angle = atan((cone->diameter / 2) / cone->height);
+	angle = atanf((cone->diameter / 2) / cone->height);
 	solve_quad_eq_cone(&eq, angle, cone, ray);
 	if (eq.discriminant < 0)
 		return ;
@@ -69,7 +69,7 @@ static void	hit_curved_surface_cone(t_conic *cone, t_ray *ray, t_hp *hp)
 			/ vec3_dot(cone->axis, cone->axis));
 	norm = vec3_unit(vec3_sub(hp_to_con_pos, proj));
 	hp->norm
-		= vec3_unit(vec3_sub(norm, vec3_scale(cone->axis, tan(angle))));
+		= vec3_unit(vec3_sub(norm, vec3_scale(cone->axis, tanf(angle))));
 	if (vec3_dot(ray->direction, hp->norm) > 0)
 		hp->norm = vec3_scale(hp->norm, -1);
 }
@@ -80,12 +80,12 @@ static void	solve_quad_eq_cone(
 	t_vec3		vec;
 
 	vec = vec3_sub(ray->origin, cone->pos);
-	eq->a = 1 - (1 + pow(tan(angle), 2))
-		* pow(vec3_dot(ray->direction, cone->axis), 2);
+	eq->a = 1 - (1 + powf(tanf(angle), 2))
+		* powf(vec3_dot(ray->direction, cone->axis), 2);
 	eq->b = (vec3_dot(ray->direction, vec)
-			- vec3_dot(ray->direction, cone->axis) * (1 + pow(tan(angle), 2))
+			- vec3_dot(ray->direction, cone->axis) * (1 + powf(tanf(angle), 2))
 			* vec3_dot(vec, cone->axis)) / eq->a;
-	eq->c = vec3_dot(vec, vec) - (1 + pow(tan(angle), 2))
-		* pow(vec3_dot(vec, cone->axis), 2);
+	eq->c = vec3_dot(vec, vec) - (1 + powf(tanf(angle), 2))
+		* powf(vec3_dot(vec, cone->axis), 2);
 	eq->discriminant = eq->b * eq->b - eq->c / eq->a;
 }
